+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # NEXT
 
+- fix TeX build date to use SOURCE_DATE_EPOCH if set, otherwise falling back to
+  the current date (previously hardcoded to Feb 8, 2025).
 - fix zealous JSON escaping
 - add `(open-base64)` command for binary data (@merv1n34k)
 - add `-stream` flag for filesystem-independent editing (@merv1n34k)